% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_absolute}
\alias{Path_absolute}
\alias{Path_resolve}
\title{Get the absolute path of the provided path.}
\usage{
Path_absolute(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A character vector containing the absolute paths.
}
\description{
This function returns the absolute path of the input path, similar to the
behavior of Python's \code{pathlib.Path.resolve()} or \code{pathlib.Path.absolute()}
methods.
}
\examples{
Path_absolute("relative/path/to/file.txt")
}
