% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_chmod}
\alias{Path_chmod}
\title{Change the mode (permissions) of a file or directory.}
\usage{
Path_chmod(paths, mode)
}
\arguments{
\item{paths}{A character vector of file paths.}

\item{mode}{The mode (permissions) to set for the file or directory.}
}
\description{
This function changes the mode (permissions) of the specified file or
directory to the provided \code{mode}.
}
\examples{
\dontrun{
Path_chmod("file.txt", "755")
}
}
