% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_cwd}
\alias{Path_cwd}
\title{Get the current working directory.}
\usage{
Path_cwd()
}
\value{
A character vector containing the current working directory.
}
\description{
This function returns the current working directory as a character vector,
similar to Python's \code{pathlib.Path.cwd()} method.
}
\examples{
Path_cwd()
}
