% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_expanduser}
\alias{Path_expanduser}
\title{Expand the tilde (~) in file paths to the user's home directory.}
\usage{
Path_expanduser(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A character vector with expanded tilde paths.
}
\description{
This function replaces the tilde (~) character with the user's home directory
path in the provided paths, similar to Python's \code{pathlib.Path.expanduser()}
method.
}
\examples{
Path_expanduser("~/Documents/file.txt")
}
