% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_glob}
\alias{Path_glob}
\title{Glob for files matching a pattern.}
\usage{
Path_glob(patterns)
}
\arguments{
\item{patterns}{A character vector of file patterns to match.}
}
\value{
A character vector containing the paths of matching files.
}
\description{
This function searches for files matching the specified patterns and returns
their paths as a character vector, similar to Python's \code{pathlib.Path.glob()}
method.
}
\examples{
Path_glob("*.txt")
}
