% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_group}
\alias{Path_group}
\title{Get the group owner of the specified files or directories.}
\usage{
Path_group(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A character vector containing the group owner of each path.
}
\description{
This function returns the group owner of the provided paths as a character
vector, similar to Python's \code{pathlib.Path.group()} method.
}
\examples{
\dontrun{
Path_group("file.txt")
}
}
