% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_home}
\alias{Path_home}
\title{Get the user's home directory.}
\usage{
Path_home()
}
\value{
A character vector containing the user's home directory path.
}
\description{
This function returns the user's home directory as a character vector,
similar to Python's \code{pathlib.Path.home()} method.
}
\examples{
\dontrun{
Path_home()
}
}
