% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_absolute}
\alias{Path_is_absolute}
\title{Check if the provided paths are absolute paths.}
\usage{
Path_is_absolute(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is an absolute path.
}
\description{
This function checks if the specified paths are absolute paths and returns a
logical vector indicating the result. An absolute path is a path that starts
from the root directory, e.g., "/path/to/file.txt", "C:\\path\\to\\file.txt"
}
\examples{
\dontrun{
Path_is_absolute("/path/to/file.txt")
}
}
