% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_block_device}
\alias{Path_is_block_device}
\title{Check if the provided paths are block devices.}
\usage{
Path_is_block_device(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a block device.
}
\description{
This function checks if the specified paths are block devices and returns a
logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_block_device()} method.
}
\examples{
\dontrun{
Path_is_block_device("file.txt")
}
}
