% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_char_device}
\alias{Path_is_char_device}
\title{Check if the provided paths are character devices.}
\usage{
Path_is_char_device(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a character device.
}
\description{
This function checks if the specified paths are character devices and returns
a logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_char_device()} method.
}
\examples{
\dontrun{
Path_is_char_device("file.txt")
}
}
