% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_dir}
\alias{Path_is_dir}
\title{Check if the provided paths are directories.}
\usage{
Path_is_dir(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a directory.
}
\description{
This function checks if the specified paths are directories and returns a
logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_dir()} method.
}
\examples{
\dontrun{
Path_is_dir("directory")
}
}
