% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_fifo}
\alias{Path_is_fifo}
\title{Check if the provided paths are FIFOs (named pipes).}
\usage{
Path_is_fifo(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a FIFO.
}
\description{
This function checks if the specified paths are FIFOs (named pipes) and
returns a logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_fifo()} method.
}
\examples{
\dontrun{
Path_is_fifo("fifo_pipe")
}
}
