% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_file}
\alias{Path_is_file}
\title{Check if the provided paths are regular files.}
\usage{
Path_is_file(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a regular file.
}
\description{
This function checks if the specified paths are regular files and returns a
logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_file()} method.
}
\examples{
\dontrun{
Path_is_file("file.txt")
}
}
