% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_socket}
\alias{Path_is_socket}
\title{Check if the provided paths are sockets.}
\usage{
Path_is_socket(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a socket.
}
\description{
This function checks if the specified paths are sockets and returns a logical
vector indicating the result. Similar to Python's \code{pathlib.Path.is_socket()}
method.
}
\examples{
\dontrun{
Path_is_socket("socket_file")
}
}
