% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_is_symlink}
\alias{Path_is_symlink}
\title{Check if the provided paths are symbolic links.}
\usage{
Path_is_symlink(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A logical vector indicating whether each path is a symbolic link.
}
\description{
This function checks if the specified paths are symbolic links and returns a
logical vector indicating the result. Similar to Python's
\code{pathlib.Path.is_symlink()} method.
}
\examples{
\dontrun{
Path_is_symlink("symbolic_link")
}
}
