% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_lstat}
\alias{Path_lstat}
\title{Get file information without following symbolic links.}
\usage{
Path_lstat(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A list containing file information for each path.
}
\description{
This function retrieves file information for the provided paths without
following symbolic links. Similar to Python's \code{pathlib.Path.lstat()} method.
}
\examples{
\dontrun{
Path_lstat("file.txt")
}
}
