% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_mkdir}
\alias{Path_mkdir}
\title{Create directories at the specified paths.}
\usage{
Path_mkdir(paths, parents = FALSE)
}
\arguments{
\item{paths}{A character vector of directory paths to create.}

\item{parents}{Logical. If TRUE, create parent directories as needed.}
}
\description{
This function creates directories at the specified paths. If the \code{parents}
argument is set to TRUE, it creates parent directories as needed. Similar to
Python's \code{pathlib.Path.mkdir()} method.
}
\examples{
\dontrun{
Path_mkdir("new_directory")
}
}
