% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_owner}
\alias{Path_owner}
\title{Get the owner of the specified files or directories.}
\usage{
Path_owner(paths)
}
\arguments{
\item{paths}{A character vector of file paths.}
}
\value{
A character vector containing the owner of each path.
}
\description{
This function returns the owner of the provided paths as a character vector,
similar to Python's \code{pathlib.Path.owner()} method.
}
\examples{
\dontrun{
Path_owner("file.txt")
}
}
