% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_read_text}
\alias{Path_read_text}
\title{Read the content of a file as text.}
\usage{
Path_read_text(path, encoding = c("unknown", "UTF-8", "latin1"))
}
\arguments{
\item{path}{A character vector specifying the path to the file.}

\item{encoding}{The character encoding to use for reading the file
("unknown", "UTF-8", or "latin1").}
}
\value{
A character vector containing the content of the file.
}
\description{
This function reads the content of a file at the specified path as text,
using the specified encoding.
}
\examples{
\dontrun{
Path_read_text("file.txt", encoding = "UTF-8")
}
}
