% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_readlink}
\alias{Path_readlink}
\title{Get the target path of a symbolic link.}
\usage{
Path_readlink(paths)
}
\arguments{
\item{paths}{A character vector of symbolic link paths.}
}
\value{
A character vector containing the target path of each symbolic link.
}
\description{
This function returns the target path of the specified symbolic links,
similar to Python's \code{pathlib.Path.readlink()} method.
}
\examples{
\dontrun{
Path_readlink("symbolic_link")
}
}
