% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_rename}
\alias{Path_rename}
\title{Rename files or directories to new paths.}
\usage{
Path_rename(paths, new_paths)
}
\arguments{
\item{paths}{A character vector of source file paths.}

\item{new_paths}{A character vector of target file paths.}
}
\description{
This function renames files or directories from their current paths to new
paths. Similar to Python's \code{pathlib.Path.rename()} method.
}
\examples{
\dontrun{
Path_rename("old_file.txt", "new_file.txt")
}
}
