% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_stat}
\alias{Path_stat}
\title{Get file information for the provided paths.}
\usage{
Path_stat(paths, ..., follow_symlinks = TRUE)
}
\arguments{
\item{paths}{A character vector of file paths.}

\item{...}{Additional arguments to pass to \code{fs::file_info()}.}

\item{follow_symlinks}{Logical. If TRUE, follow symbolic links to obtain file
information.}
}
\value{
A list containing file information for each path.
}
\description{
This function retrieves file information for the provided paths, allowing an
optional argument \code{follow_symlinks} to specify whether symbolic links should
be followed when obtaining file information. Similar to Python's
\code{pathlib.Path.stat()} method.
}
\examples{
\dontrun{
Path_stat("file.txt", follow_symlinks = TRUE)
}
}
