% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_symlink_to}
\alias{Path_symlink_to}
\title{Create symbolic links from the paths to the target paths.}
\usage{
Path_symlink_to(paths, targets)
}
\arguments{
\item{paths}{A character vector of source file paths.}

\item{targets}{A character vector of target file paths.}
}
\description{
This function creates symbolic links from the specified paths to the target
paths. Similar to Python's \code{pathlib.Path.symlink_to()} method.
}
\examples{
\dontrun{
Path_symlink_to("source_file.txt", "target_file.txt")
}
}
