% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_touch}
\alias{Path_touch}
\title{Touch files at the specified paths.}
\usage{
Path_touch(paths)
}
\arguments{
\item{paths}{A character vector of file paths to touch.}
}
\description{
This function creates the files if they do not exist and updates the access
and modification times if they do exist. Similar to the Unix \code{touch} command.
}
\examples{
\dontrun{
Path_touch("file.txt")
}
}
