% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_unlink}
\alias{Path_unlink}
\title{Unlink (delete) files at the specified paths.}
\usage{
Path_unlink(paths, missing_ok = FALSE)
}
\arguments{
\item{paths}{A character vector of file paths to unlink.}

\item{missing_ok}{Logical. If TRUE, allow unlinking of non-existent files.}
}
\description{
This function unlinks (deletes) files at the specified paths. It does not
remove directories. If the \code{missing_ok} argument is set to FALSE, it checks
if the files exist before attempting to unlink them. Similar to Python's
\code{pathlib.Path.unlink()} method.
}
\examples{
\dontrun{
Path_unlink("file.txt")
}
}
