% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathlib.R
\name{Path_write_text}
\alias{Path_write_text}
\title{Write text to a file at the specified path.}
\usage{
Path_write_text(path, text)
}
\arguments{
\item{path}{A character vector specifying the path to the file.}

\item{text}{The text to write to the file.}
}
\description{
This function writes the provided text to a file at the specified path.
}
\examples{
\dontrun{
Path_write_text("file.txt", "Hello, World!")
}
}
